+++ src/action/notSalesforceSetup/notSalesforceSetup.js
@@ -11,7 +11,7 @@ import {
 	WHAT_GET_BROWSER_TAB,
 } from "../../core/constants.js";
 import { getSettings, sendExtensionMessage } from "../../core/functions.js";
-import ensureTranslatorAvailability from "../../core/translator.js";
+import { ensureTranslatorAvailability } from "../../core/translator.js";
 import "../themeHandler.js";
 
 /**